backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
elasticsearch==8.9.0
python-dotenv==1.0.0
nanoid==2.0.0
requests==2.31.0
pytest==7.4.0